refactor(history): combine function history&detail, cleanCode, fix some bug
