+++ src/routes/historyRoutes.js
@@ -1,10 +1,9 @@
 import { Router } from 'express';
-import { getDetailHistoryTransaction, getHistoryTransaction } from '../controllers/historyController.js';
+import { getHistoryTransactionAndDetail } from '../controllers/historyController.js';
 import { isAuthenticate } from "../middlewares/auth.js";
 
 const historyRoutes = new Router();
 
-historyRoutes.get('/', isAuthenticate, getHistoryTransaction);
-historyRoutes.get('/detail/:transactionId', isAuthenticate,getDetailHistoryTransaction);
+historyRoutes.get('/:userId',isAuthenticate, getHistoryTransactionAndDetail);
 
 export default historyRoutes;
